+++ include/generic.h
@@ -391,7 +391,7 @@ size_t node_index(int x,int y,int z,int t);
 size_t num_sites(int node);
 const int *get_logical_dimensions(void);
 const int *get_logical_coordinate(void);
-void get_coords(int coords[], int node, int index);
+void get_coords(int coords[], int node, size_t index);
 
 /* make_lattice.c */
 void make_lattice(void);
